+++ changelog.md
@@ -18,6 +18,7 @@
 * [3520](https://github.com/zeta-chain/node/pull/3520) - integrate SPL withdraw and call
 * [3527](https://github.com/zeta-chain/node/pull/3527) - integrate SOL/SPL withdraw and call revert
 * [3522](https://github.com/zeta-chain/node/pull/3522) - add `MsgDisableFastConfirmation` to disable fast confirmation. This message can be triggered by the emergency policy.
+* [3551](https://github.com/zeta-chain/node/pull/3551) - support for EVM chain and Bitcoin chain inbound fast confirmation
 
 ### Refactor